Update client to support V3
Still some tests to be written.

Took the opportunity to refactor a bit. It won't strictly be backwards
compatible with older client versions since the exceptions changed
namespace and names. I did this because it was frankly a bit of a mess
before. Also removed some now unused internal classes which should not
have nbeen directly used by the client's clients. Other than that it
should play nicely with old code.